Revert "float version"
This reverts commit d898e2bd36043367dd6ce510809806dfdf701a0d.

+++ Temperature.c
@@ -4,7 +4,7 @@
 当fahr = 0, 20, ··· , 300 时,分别打印华氏温度和摄氏温度对照表
 */
 main(){
-    float fahr, celsius;
+    int fahr, celsius;
     int lower, upper, step;
 
     lower = 0; // 温度表的下限
@@ -14,7 +14,7 @@ main(){
     fahr = lower;
     while(fahr <= upper){
         celsius = 5 * (fahr - 32) / 9;
-        printf("%3.0f\t%6.1f\n", fahr, celsius );
+        printf("%d\t%d\n", fahr, celsius );
         fahr = fahr + step;
     }
 }
